Removed android_old target and began update to sdk install
