+++ buildozer/__init__.py
@@ -155,7 +155,6 @@ class Buildozer(object):
     def set_target(self, target):
         '''Set the target to use (one of buildozer.targets, such as "android")
         '''
-        target = self.translate_target(target)
         self.targetname = target
         m = __import__('buildozer.targets.{0}'.format(target),
                 fromlist=['buildozer'])
@@ -919,23 +918,6 @@ class Buildozer(object):
     # command line invocation
     #
 
-    def translate_target(self, target, inverse=False):
-        # FIXME at some point, refactor to remove completely android old toolchain
-        if inverse:
-            if target == "android":
-                target = "android_old"
-            elif target == "android_new":
-                target = "android"
-        else:
-            if target == "android":
-                target = "android_new"
-            elif target == "android_new":
-                self.error("ERROR: The target android_new is now android")
-                exit(1)
-            elif target == "android_old":
-                target = "android"
-        return target
-
     def targets(self):
         for fn in listdir(join(dirname(__file__), 'targets')):
             if fn.startswith('.') or fn.startswith('__'):
@@ -946,7 +928,7 @@ class Buildozer(object):
             try:
                 m = __import__('buildozer.targets.{0}'.format(target),
                         fromlist=['buildozer'])
-                yield self.translate_target(target, inverse=True), m
+                yield target, m
             except NotImplementedError:
                 pass
             except:
@@ -1054,7 +1036,7 @@ class Buildozer(object):
 
         # maybe it's a target?
         targets = [x[0] for x in self.targets()]
-        if self.translate_target(command, inverse=True) not in targets:
+        if command not in targets:
             print('Unknown command/target {}'.format(self.translate_target(command, inverse=True)))
             exit(1)
